refactor: move `crop` out of jsfeat folder
It doesn't use any of the jsfeat data structures.
